pyserverforindy/requirements.txt
async-generator==1.10
atomicwrites==1.2.1
attrs==18.2.0
base58==1.0.3
grpcio==1.26.0
grpcio-tools==1.26.0
more-itertools==5.0.0
pluggy==0.7.1
protobuf==3.6.1
py==1.7.0
pytest==3.6.4
pytest-asyncio==0.9.0
python3-indy==1.7.0
six==1.12.0